refact : 네비게이션 바 렌더링 요소 분리 nav -> nav > navCell
